undo changes except KeychainDisabled check in deleteKeychainItems

+++ Source/Authentication/ZMPersistentCookieStorage.m
@@ -29,6 +29,7 @@
 
 static NSString * const LegacyAccountName = @"User";
 static NSString* ZMLogTag ZM_UNUSED = ZMT_LOG_TAG_NETWORK;
+static BOOL KeychainDisabled = NO;
 static NSMutableDictionary *NonPersistedPassword;
 static NSHTTPCookieAcceptPolicy cookiesPolicy = NSHTTPCookieAcceptPolicyAlways;
 
@@ -89,6 +90,11 @@ static dispatch_queue_t isolationQueue()
 
 #pragma mark - Public API
 
++ (void)setDoNotPersistToKeychain:(BOOL)disabled;
+{
+    KeychainDisabled = disabled;
+}
+
 - (NSData *)authenticationCookieData;
 {
     NSData *result = nil;
@@ -130,6 +136,10 @@ static dispatch_queue_t isolationQueue()
     dispatch_sync(isolationQueue(), ^{
         NonPersistedPassword[self.cookieKey] = nil;
 
+//        if (KeychainDisabled) {
+//            return;
+//        }
+
         [ZMKeychain deleteAllKeychainItemsWithAccountName:self.accountName];
     });
 }
@@ -138,6 +148,10 @@ static dispatch_queue_t isolationQueue()
 {
     dispatch_sync(isolationQueue(), ^{
         NonPersistedPassword = nil;
+
+        if (KeychainDisabled) {
+            return;
+        }
 
         [ZMKeychain deleteAllKeychainItems];
     });
@@ -157,6 +171,11 @@ static dispatch_queue_t isolationQueue()
         NSData *password = NonPersistedPassword[self.cookieKey];
         BOOL const fetchFromKeychain = (password == nil);
         *passwordP = (password == (id) [NSNull null]) ? nil : password;
+        
+        if (KeychainDisabled) {
+            success = YES;
+            return;
+        }
         
         if (fetchFromKeychain) {
             id result = nil;
@@ -204,6 +223,10 @@ static dispatch_queue_t isolationQueue()
         
         [self addNonPersistedPassword:password];
         
+        if (KeychainDisabled) {
+            success = YES;
+            return;
+        }
         success = [ZMKeychain setData:password forAccount:self.accountName];
     });
     return success;
@@ -218,6 +241,11 @@ static dispatch_queue_t isolationQueue()
                         (NonPersistedPassword[self.cookieKey] != [NSNull null]));
         if (hasItem) {
             NonPersistedPassword[self.cookieKey] = password ?: [NSNull null];
+        }
+        
+        if (KeychainDisabled) {
+            success = hasItem;
+            return;
         }
         
         success = [ZMKeychain setData:password forAccount:self.accountName];
@@ -242,6 +270,10 @@ static dispatch_queue_t isolationQueue()
     dispatch_sync(isolationQueue(), ^{
         
         [NonPersistedPassword removeObjectForKey:self.cookieKey];
+        
+        if (KeychainDisabled) {
+            return;
+        }
         
         [ZMKeychain deleteAllKeychainItemsWithAccountName:self.accountName];
     });
